Refactor mixer
