+++ src/meson.build
@@ -64,28 +64,9 @@ subdir('scrap')
 # optional modules
 subdir('font')
 
-# TODO: support SDL3
+subdir('mixer')
+
 if sdl_api != 3
-
-if sdl_mixer_dep.found()
-    mixer = py.extension_module(
-        'mixer',
-        'mixer.c',
-        c_args: warnings_error,
-        dependencies: pg_base_deps + sdl_mixer_dep,
-        install: true,
-        subdir: pg,
-    )
-
-    mixer_music = py.extension_module(
-        'mixer_music',
-        'music.c',
-        c_args: warnings_error,
-        dependencies: pg_base_deps + sdl_mixer_dep,
-        install: true,
-        subdir: pg,
-    )
-endif
 
 if freetype_dep.found()
     _freetype = py.extension_module(
