+++ src/mixer/music_doc.h
